refactor, move database functions to models
